Fixes defect #35: Duplicate folders when scrolling ListBox.
Description: This defect apply when "ListBox" use "VirtualizingStackPanel" which "VirtualizingStackPanel.VirtualizationMode" property is set to "Recycle" (one of the .NET 3.5 SP1 performance improvements). Setting  "VirtualizingStackPanel.VirtualizationMode" property to defaul "Standart" value fix this, but scrolling performans in "ListBox" with more than 100 items is worse then with "Recycly" mode.

Note: Setting "ScrollViewer.CanContentScroll" to "False" (to achieve smooth scrolling of content in "ListBox") for the internal scroll viewer of this "ListBox" will disable vistualization of "VirtualizingStackPanel" at all, even if the "VirtualizingStackPanel.IsVirtualizing" is set to "True" manualy (by default it is "True" anyway)!